requirements.txt
psycopg2-binary
python-dotenv
sentence-transformers
numpy
huggingface_hub[hf_xet]==0.36
langgraph
langchain
langchain-openai
fastapi
uvicorn
streamlit
requests
transformers
huggingface-hub
torch
accelerate
chromadb